% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEncoder.r
\name{buildEncoder}
\alias{buildEncoder}
\title{Build autoencoder}
\usage{
buildEncoder(trainData, valData, encoderSettings, imageProcessingSettings,
  outputFolder)
}
\arguments{
\item{outputFolder}{}

\item{x}{An object}
}
\description{
Build autoencoder
}
