% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEncoder.r
\name{fit2DConvAutoencoder}
\alias{fit2DConvAutoencoder}
\title{Fit 2-dimensional autoencoder}
\usage{
fit2DConvAutoencoder(trainData, valProp, epochs = epochs,
  batchSize = batchSize, poolingLayerNum = poolingLayerNum,
  kernelSize = kernelSize, poolSize = poolSize, optimizer = "adadelta",
  loss = "binary_crossentropy",
  imageProcessingSettings = imageProcessingSettings)
}
\description{
Fit 2-dimensional autoencoder
}
