% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEncoder.r
\name{fitVanillaAutoencoder}
\alias{fitVanillaAutoencoder}
\title{Fit Vanilla autoencoder}
\usage{
fitVanillaAutoencoder(trainData, valProp, epochs = epochs,
  batchSize = batchSize, latentDim = latentDim, optimizer = "adadelta",
  loss = "binary_crossentropy",
  imageProcessingSettings = imageProcessingSettings)
}
\description{
Fit Vanilla autoencoder
}
