% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.r
\name{meltDim}
\alias{meltDim}
\title{melt the dimension of the datda}
\usage{
meltDim(x, convolution = F, imageProcessingSettings)
}
\arguments{
\item{x}{An array object}

\item{imageProcessingSettings}{class of imageProcessingSettings}
}
\description{
melt the dimension of the datda
}
\details{
This function is for preprocessing fo the data, including normalization, resizing, and dimension concat
}
