% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.r
\name{preProcessing}
\alias{preProcessing}
\title{Pre-processing function for the data}
\usage{
preProcessing(x = NULL, imageProcessingSettings = imageProcessingSettings)
}
\arguments{
\item{x}{An object of data}

\item{imageProcessingSettings}{class of imageProcessingSettings}
}
\description{
Pre-processing function for the data
}
\details{
This function is for preprocessing fo the data, including normalization, resizing, and dimension concat
}
