% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.r
\name{reconDim}
\alias{reconDim}
\title{Reconstruction of the dimenion of the data}
\usage{
reconDim(x, convolution = F, imageProcessingSettings)
}
\arguments{
\item{x}{An object of data}

\item{imageProcessingSettings}{imageProcessingSettings}
}
\description{
Reconstruction of the dimenion of the data
}
\details{
This function is for preprocessing fo the data, including normalization, resizing, and dimension concat
}
