% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.r
\name{reverseProcessing}
\alias{reverseProcessing}
\title{Reverse-processing function for the data}
\usage{
reverseProcessing(x)
}
\arguments{
\item{x}{An object of data}

\item{normalization}{method for normalization. 'MinMaxNorm' is min-max normalization}

\item{dimConcat}{dimConcat indicates the index dimension. dimConcat should be NULL or an integer. if dimConcat is not null, the array will be concatenated along this specified dimension}
}
\description{
Reverse-processing function for the data
}
\details{
This function is for preprocessing fo the data, including normalization, resizing, and dimension concat
}
