% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAutoencoder.r
\name{set2DConvAutoencoder}
\alias{set2DConvAutoencoder}
\title{Creating setting for 2-dimension convolutional autoencoder}
\usage{
set2DConvAutoencoder(valProp = 0.3, epochs = 10, batchSize = batchSize,
  poolingLayerNum = 3, kernelSize = 3, poolSize = 2,
  optimizer = "adadelta", loss = "binary_crossentropy",
  imageProcessingSettings = imageProcessingSettings)
}
\arguments{
\item{latentDim}{}
}
\description{
Creating setting for 2-dimension convolutional autoencoder
}
