% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.r
\name{SetImageProcessing}
\alias{SetImageProcessing}
\title{Data processing setting}
\usage{
SetImageProcessing(normalization = NULL, maxLimit = NULL, minLimit = NULL,
  width = 28, height = 28, roiWidth = NULL, roiHeight = NULL,
  channelDim = NULL, indexDim = NULL)
}
\arguments{
\item{normalization}{method for normalization. 'MinMaxNorm' is min-max normalization}

\item{maxLimit}{maximum value for all values of the data}

\item{roiWidth}{Region of interest, width. default seq(width)}

\item{roiHeight}{Region of interest, height. default seq(height)}

\item{channelDim}{An additional dimension for the channel of the image. It there is no channel, it should be NULL}

\item{indexDim}{indexDim indicates the index dimension. indexDim should be NULL or an integer. if indexDim is not null, the array will be melted along this specified dimension}
}
\description{
Data processing setting
}
\details{
This function is for preprocessing fo the data, including normalization, resizing, and dimension concat
}
