% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAutoencoder.r
\name{setVanillaAutoencoder}
\alias{setVanillaAutoencoder}
\title{Creating setting for Vanilla autoencoder}
\usage{
setVanillaAutoencoder(valProp = 0.3, epochs = 10, batchSize = batchSize,
  latentDim = latentDim, optimizer = "adadelta",
  loss = "binary_crossentropy",
  imageProcessingSettings = imageProcessingSettings)
}
\arguments{
\item{latentDim}{}
}
\description{
Creating setting for Vanilla autoencoder
}
